[FE-1192] Made ReportsListModal into a shared component
