+++ src/components/reportBuilder/ReportsListModalTabs.js
@@ -1,25 +1,18 @@
-import React, { useState } from 'react';
-import { connect, useDispatch } from 'react-redux';
+import React from 'react';
 import { MoreHoriz } from '@sparkpost/matchbox-icons';
 import { TableCollection, Subaccount } from 'src/components';
 import {
   ActionList,
   Box,
   Button,
-  Modal,
   Popover,
   Radio,
   ScreenReaderOnly,
   Table,
-  Tabs,
   Tag,
 } from 'src/components/matchbox';
 import { formatDateTime } from 'src/helpers/date';
 import { ButtonLink, PageLink } from 'src/components/links';
-import { selectCondition } from 'src/selectors/accessConditionState';
-import { isAccountUiOptionSet } from 'src/helpers/conditions/account';
-import { updateUserUIOptions } from 'src/actions/currentUser';
-import { showAlert } from 'src/actions/globalAlert';
 
 const allReportsColumns = [
   { label: 'Name', sortKey: 'name' },
@@ -69,32 +62,28 @@ const Actions = ({ id, handleDelete, handleEdit, reportType, report, ...rest }) 
   );
 };
 
-export function ReportsListModal({
+export const MyReportsTab = ({
   reports,
-  open,
-  onClose,
   currentUser,
+  onDashboard,
+  handleRadioChange,
+  selectedReportId,
+  handleReportChangeAndClose,
+  isScheduledReportsEnabled,
   handleDelete,
   handleEdit,
-  isScheduledReportsEnabled,
-  onDashboard,
-  handleReportChange,
-}) {
-  const handleReportChangeAndClose = report => {
-    handleReportChange(report);
-    onClose();
-  };
-
+}) => {
   const myReports = reports.filter(({ creator }) => creator === currentUser);
+  const getMyReportColumns = () => {
+    if (onDashboard)
+      return [
+        {},
+        { label: 'Name', sortKey: 'name' },
+        { label: 'Last Modification', sortKey: 'modified' },
+      ];
 
-  const dispatch = useDispatch();
-
-  const [tabIndex, setTabIndex] = useState(0);
-
-  const [selectedReportId, setSelectedReportId] = useState(null);
-
-  const handleRadioChange = id => setSelectedReportId(id);
-
+    return myReportsColumns;
+  };
   const myReportsRows = report => {
     const { name, modified, isLast, id } = report;
     if (onDashboard)
@@ -130,18 +119,34 @@ export function ReportsListModal({
       />,
     ];
   };
+  return (
+    <TableCollection
+      rows={myReports}
+      columns={getMyReportColumns()}
+      getRowData={myReportsRows}
+      wrapperComponent={Table}
+      filterBox={{
+        label: '',
+        show: true,
+        itemToStringKeys: ['name', 'modified'],
+        exampleModifiers: ['name', 'modified'],
+        maxWidth: '1250',
+        wrapper: FilterBoxWrapper,
+      }}
+    />
+  );
+};
 
-  const getMyReportColumns = () => {
-    if (onDashboard)
-      return [
-        {},
-        { label: 'Name', sortKey: 'name' },
-        { label: 'Last Modification', sortKey: 'modified' },
-      ];
-
-    return myReportsColumns;
-  };
-
+export const AllReportsTab = ({
+  reports,
+  onDashboard,
+  handleRadioChange,
+  selectedReportId,
+  handleReportChangeAndClose,
+  isScheduledReportsEnabled,
+  handleDelete,
+  handleEdit,
+}) => {
   const getColumnsForAllReports = () => {
     if (onDashboard)
       return [
@@ -203,46 +208,7 @@ export function ReportsListModal({
       action,
     ];
   };
-
-  const onSubmit = () => {
-    dispatch(updateUserUIOptions({ pinned_report: selectedReportId })).then(() => {
-      dispatch(
-        showAlert({
-          type: 'success',
-          message: 'Pinned Report updated',
-        }),
-      );
-    });
-
-    onClose();
-  };
-
-  const ModalContentContainer = ({ children }) => {
-    if (!onDashboard) return <>{children}</>;
-
-    return (
-      <Radio.Group label="reportList" labelHidden>
-        {children}
-      </Radio.Group>
-    );
-  };
-
-  const TABS = [
-    <TableCollection
-      rows={myReports}
-      columns={getMyReportColumns()}
-      getRowData={myReportsRows}
-      wrapperComponent={Table}
-      filterBox={{
-        label: '',
-        show: true,
-        itemToStringKeys: ['name', 'modified'],
-        exampleModifiers: ['name', 'modified'],
-        maxWidth: '1250',
-        wrapper: FilterBoxWrapper,
-      }}
-    />,
-
+  return (
     <TableCollection
       rows={reports}
       columns={getColumnsForAllReports()}
@@ -256,51 +222,6 @@ export function ReportsListModal({
         maxWidth: '1250',
         wrapper: FilterBoxWrapper,
       }}
-    />,
-  ];
-
-  return (
-    <Modal open={open} onClose={onClose} showCloseButton maxWidth="1300">
-      <Modal.Header>{onDashboard ? 'Change Report' : 'Saved Reports'}</Modal.Header>
-      <Modal.Content>
-        <ModalContentContainer>
-          <Tabs
-            tabs={[
-              { content: 'My Reports', onClick: () => setTabIndex(0) },
-              { content: 'All Reports', onClick: () => setTabIndex(1) },
-            ]}
-            fitted
-            selected={tabIndex}
-          />
-
-          {TABS[tabIndex]}
-        </ModalContentContainer>
-      </Modal.Content>
-      {onDashboard && (
-        <Modal.Footer>
-          <Button
-            variant="primary"
-            loadingLabel="Loading"
-            onClick={onSubmit}
-            disabled={!selectedReportId}
-          >
-            Change Report
-          </Button>
-          <Button variant="secondary" onClick={onClose}>
-            Cancel
-          </Button>
-        </Modal.Footer>
-      )}
-    </Modal>
+    />
   );
-}
-
-const mapStateToProps = state => {
-  return {
-    currentUser: state.currentUser.username,
-    isScheduledReportsEnabled: selectCondition(isAccountUiOptionSet('allow_scheduled_reports'))(
-      state,
-    ),
-  };
 };
-export default connect(mapStateToProps)(ReportsListModal);
