+++ src/components/reportBuilder/index.js
@@ -1,0 +1,1 @@
+export { default as ReportsListModal } from './ReportsListModal';
