+++ src/pages/reportBuilder/components/SavedReportsSection/SavedReportsSection.js
@@ -8,9 +8,8 @@ import { AccessTime, Edit, FolderOpen, Save } from '@sparkpost/matchbox-icons';
 import SaveReportModal from './SaveReportModal';
 import { deleteReport, getReports } from 'src/actions/reports';
 import useModal from 'src/hooks/useModal';
-import ReportsListModal from './ReportsListModal';
 import ScheduledReportsModal from './ScheduledReportsModal';
-import { DeleteModal } from 'src/components/modals';
+import { DeleteModal, ReportsListModal } from 'src/components';
 import { showAlert } from 'src/actions/globalAlert';
 import { selectCondition } from 'src/selectors/accessConditionState';
 import { isAccountUiOptionSet } from 'src/helpers/conditions/account';
